% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ae_aedecod.R
\name{check_ae_aedecod}
\alias{check_ae_aedecod}
\title{Check for missing AEDECOD values}
\usage{
check_ae_aedecod(AE, preproc = identity, ...)
}
\arguments{
\item{AE}{Adverse Events SDTM dataset with variables USUBJID, AETERM, AEDECOD}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the test failed
}
\description{
This check looks for missing AEDECOD values
}
\examples{

# test with sample data

AE <- data.frame(
 USUBJID = 1:5,
 DOMAIN = c(rep("AE", 5)),
 AESEQ = 1:5,
 AESTDTC = 1:5,
 AETERM = 1:5,
 AEDECOD = 1:5,
 AESPID = c("/F:AEDE-D:47155-R:1/L:5/AT:INITIALEXTREME",
 "/F:AEDE-D:43355-R:2/L:5/AT:INITIALEXTREME",
 "/F:AEDE-D:43155-R:3/L:5/AT:INITIALEXTREME",
 "/F:TOXICITY_OTH-D:119232-R:97/L:5/AT:INITIALEXTREME",
 "/F:TOXICITY_OTH-D:119232-R:90"),
 stringsAsFactors = FALSE
)

check_ae_aedecod(AE)

AE$AEDECOD[1] = NA
AE$AEDECOD[2] = "NA"
AE$AEDECOD[3:5] = ""
check_ae_aedecod(AE)
check_ae_aedecod(AE,preproc=roche_derive_rave_row)

AE$AEDECOD <- NULL
check_ae_aedecod(AE)




}
\author{
Yinghui Miao, Stella Banjo(HackR 2021)
}
