% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roche_functions.R
\name{roche_derive_rave_row}
\alias{roche_derive_rave_row}
\title{Utility function to obtain Rave row}
\usage{
roche_derive_rave_row(dts, domains = c("ae", "cm"))
}
\arguments{
\item{dts}{dataset e.g. AE}

\item{domains}{domains you wish to identify a SPID variable from}
}
\value{
dataset with rave row number
}
\description{
This function derives the Rave row number from XXSPID
}
\examples{

AE <- data.frame(
  STUDY = c(rep("GO29537", 6)),
  DOMAIN = c(rep("AE", 6)),
  USUBJID = c(rep("GO29537-278300-30221", 6)),
  AESEQ = c(1, 2, 3, 4, 5, 6),
  AETERM = c("ANEMIA", "ANEMIA", "ANOREXIA", "CONSTIPATION", "DIARRHEA", "FATIGUE"),
  AEDECOD = c("ANAEMIA", "ANAEMIA", "DECREASED APPETITE", "CONSTIPATION", "DIARRHOEA", "FATIGUE"),
  AESPID = c("/F:AEDE-D:22028232-R:13/L:13/AT:INITIALEXTREME",
             "/F:AEDE-D:22028232-R:16/L:16/AT:INITIALEXTREME",
             "/F:AEDE-D:22028232-R:2/L:2/AT:INITIALEXTREME",
             "/F:AEDE-D:22028232-R:19/L:19/AT:INITIALEXTREME",
             "/F:AEDE-D:22028232-R:5/L:5/AT:INITIALEXTREME",
             "/F:AEDE-D:22028232-R:20/L:20/AT:INITIALEXTREME"),
  AESTDTC = c("2015-08-12", "2015-09-02", "2015-07-02", "2015-08-19", "2015-07-17", "2015-08-12")
)

roche_derive_rave_row(AE)


RS <- data.frame(
   USUBJID = c("1"),
   RSSPID = c("/F:EDT_RSP", "/F:MRSP1-D:23456-R:0", "/F:RRSP1-D:23456-R:0")
   )

roche_derive_rave_row(RS)

SS <- data.frame(
   USUBJID = c("1"),
   SSSPID = c("/F:EDT_SS", "/F:LTFU-D:23456-R:0", "/F:LTFU-D:23456-R:0", "123sdf",
   "/F:AEDE-D:22028232-R:20/L:20/AT:INITIALEXTREME")
   )

roche_derive_rave_row(SS)


}
\author{
Stella Banjo (HackR 2021)
}
